chore: SequentialTaskSetWithFailure class
